+++ packages/vibrant-components/src/lib/OutlinedButton/OutlinedButton.tsx
@@ -1,5 +1,4 @@
 import { Box, Text } from '@vibrant-ui/core';
-import { Icon } from '@vibrant-ui/icons';
 import { HStack } from '../HStack';
 import { Pressable } from '../Pressable';
 import { Spinner } from '../Spinner';
@@ -12,7 +11,7 @@ export const OutlinedButton = withOutlinedButtonVariation(
     backgroundColor,
     borderColor,
     IconComponent,
-    disclosure,
+    DisclosureIconComponent,
     loading,
     disabled,
     typography,
@@ -20,6 +19,7 @@ export const OutlinedButton = withOutlinedButtonVariation(
     iconSize,
     disclosureSize,
     children,
+    contentsSpacing,
     ...restProps
   }) => (
     <Pressable
@@ -32,20 +32,33 @@ export const OutlinedButton = withOutlinedButtonVariation(
       backgroundColor={backgroundColor}
       borderColor={borderColor}
       borderWidth={1}
+      borderStyle="solid"
       borderRadiusLevel={1}
       disabled={loading || disabled}
     >
       <>
-        <HStack as="span" alignVertical="center" alignHorizontal="center" opacity={loading ? 0 : 1}>
+        <HStack
+          as="span"
+          alignVertical="center"
+          alignHorizontal="center"
+          opacity={loading ? 0 : 1}
+          spacing={contentsSpacing}
+        >
           {IconComponent && (
             <Box as="span" mx={2}>
               <IconComponent size={iconSize} />
             </Box>
           )}
-          <Text typography={typography} fontWeight="bold" ml={4} mr={disclosure ? 6 : 4}>
-            {children}
-          </Text>
-          {disclosure && <Icon.ArrowTriangleDown.Fill size={disclosureSize} />}
+          {Boolean(children) && (
+            <Text typography={typography} fontWeight="bold" mx={4}>
+              {children}
+            </Text>
+          )}
+          {DisclosureIconComponent && (
+            <Box as="span" ml={2}>
+              <DisclosureIconComponent size={disclosureSize} />
+            </Box>
+          )}
         </HStack>
         {loading && (
           <HStack