+++ packages/vibrant-components/src/lib/OutlinedButton/OutlinedButtonProps.ts
@@ -2,6 +2,7 @@ import type { Ref } from 'react';
 import type { ResponsiveValue, TextChildren } from '@vibrant-ui/core';
 import { propVariant, withVariation } from '@vibrant-ui/core';
 import type { IconComponent, IconProps } from '@vibrant-ui/icons';
+import { Icon } from '@vibrant-ui/icons';
 import type { PressableProps } from '../Pressable';
 
 export type OutlinedButtonProps = {
@@ -10,6 +11,7 @@ export type OutlinedButtonProps = {
   type?: PressableProps['buttonType'];
   IconComponent?: IconComponent<IconProps, 'Fill' | 'Regular'>;
   full?: ResponsiveValue<boolean>;
+  active?: boolean;
   disclosure?: boolean;
   disabled?: boolean;
   loading?: boolean;
@@ -33,14 +35,16 @@ export const withOutlinedButtonVariation = withVariation<OutlinedButtonProps>('O
         spinnerSize: 'sm',
         iconSize: 16,
         disclosureSize: 16,
+        contentsSpacing: 0,
       },
       md: {
         typography: 'body2',
-        py: 8,
+        py: 9,
         px: 9,
         spinnerSize: 'md',
         iconSize: 18,
         disclosureSize: 16,
+        contentsSpacing: 0,
       },
       lg: {
         typography: 'body2',
@@ -49,6 +53,7 @@ export const withOutlinedButtonVariation = withVariation<OutlinedButtonProps>('O
         spinnerSize: 'md',
         iconSize: 18,
         disclosureSize: 16,
+        contentsSpacing: 0,
       },
       xl: {
         typography: 'body1',
@@ -57,6 +62,7 @@ export const withOutlinedButtonVariation = withVariation<OutlinedButtonProps>('O
         spinnerSize: 'md',
         iconSize: 20,
         disclosureSize: 18,
+        contentsSpacing: 2,
       },
     } as const,
   }),
@@ -94,6 +100,29 @@ export const withOutlinedButtonVariation = withVariation<OutlinedButtonProps>('O
       false: {
         width: 'auto',
       },
+    },
+  }),
+  propVariant({
+    props: [
+      {
+        name: 'disclosure',
+      },
+      {
+        name: 'active',
+      },
+    ],
+    variants: ({ disclosure, active }) => {
+      if (!disclosure) {
+        return {
+          DisclosureIconComponent: null,
+        };
+      }
+
+      if (active) {
+        return { DisclosureIconComponent: Icon.ArrowTriangleUp.Fill };
+      }
+
+      return { DisclosureIconComponent: Icon.ArrowTriangleDown.Fill };
     },
   })
 );