+++ packages/vibrant-components/CHANGELOG.md
@@ -2,6 +2,35 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.26.1](https://github.com/pedaling/opensource/compare/vibrant-components-0.26.0...vibrant-components-0.26.1) (2022-11-09)
+
+
+### Features
+
+* Callout 이 string 말고도 받을 수 있도록 수정한다 ([#418](https://github.com/pedaling/opensource/issues/418)) ([1942479](https://github.com/pedaling/opensource/commit/194247962b06c870d9e42322a27830b505909da1))
+* change button option field name title to text in ModalBottomSheet (API Changed) ([#417](https://github.com/pedaling/opensource/issues/417)) ([6d7d39a](https://github.com/pedaling/opensource/commit/6d7d39aed0cbfd8edd3df20c548546a482fdbcd5))
+
+
+### Bug Fixes
+
+* set borderStyle solid in OutlinedButton ([#416](https://github.com/pedaling/opensource/issues/416)) ([720d4c5](https://github.com/pedaling/opensource/commit/720d4c52a1bb9938d44fa3058483d314855ae13c))
+
+## [0.26.0](https://github.com/pedaling/opensource/compare/vibrant-components-0.25.6...vibrant-components-0.26.0) (2022-11-08)
+
+### [0.25.5](https://github.com/pedaling/opensource/compare/vibrant-components-0.25.4...vibrant-components-0.25.5) (2022-11-07)
+
+
+### Features
+
+* export modal bottom sheet props ([#408](https://github.com/pedaling/opensource/issues/408)) ([0bf542c](https://github.com/pedaling/opensource/commit/0bf542ca9faae6cba6c5a809302470557ba3ae62))
+* GhostButton QA ([#411](https://github.com/pedaling/opensource/issues/411)) ([09a1e7f](https://github.com/pedaling/opensource/commit/09a1e7f28234bc6682732a25d3b2a46557d104af))
+
+
+### Bug Fixes
+
+* add alert and scroll movement in TextField ([#413](https://github.com/pedaling/opensource/issues/413)) ([122d985](https://github.com/pedaling/opensource/commit/122d9859054483db5542ec5922f83511fbb2ab4b))
+* OutlinedButton QA ([#410](https://github.com/pedaling/opensource/issues/410)) ([0615cfb](https://github.com/pedaling/opensource/commit/0615cfb5151e6ce9fd93db310c8c3d10b4c939cd))
+
 ### [0.25.4](https://github.com/pedaling/opensource/compare/vibrant-components-0.25.3...vibrant-components-0.25.4) (2022-11-03)